requirements.txt
pyserial==3.5
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
loguru
fastapi 
uvicorn
httpx
